requirements.txt
gigachain_community
gigachain
faiss-cpu
streamlit
python-dotenv